add tests + general function in helper file
