+++ smithy-swift-codegen/src/test/kotlin/serde/awsjson11/OutputResponseDeserializerTests.kt
@@ -98,6 +98,18 @@ class OutputResponseDeserializerTests {
                         let messageDecoder: ClientRuntime.MessageDecoder? = nil
                         let decoderStream = ClientRuntime.EventStream.DefaultMessageDecoderStream<EventStream>(stream: stream, messageDecoder: messageDecoder, responseDecoder: responseDecoder)
                         self.eventStream = decoderStream.toAsyncStream()
+                        if let initialDataWithoutHttp = await messageDecoder.awaitInitialResponse() {
+                            let decoder = JSONDecoder()
+                            do {
+                                let response = try decoder.decode([String: String].self, from: initialDataWithoutHttp)
+                                self.metadata = response["metadata"].map { value in KinesisClientTypes.Tag(value: value) }
+                            } catch {
+                                print("Error decoding JSON: \(error)")
+                                self.metadata = nil
+                            }
+                        } else {
+                            self.metadata = nil
+                        }
                     } else {
                         self.eventStream = nil
                     }
