+++ smithy-swift-codegen/src/test/resources/serde/awsjson11/awsjson-output-response-deserializer.smithy
@@ -51,6 +51,10 @@ operation EventStreaming {
 }
 
 structure EventStreamingOutput {
+    @httpLabel
+    @required
+    metadata: String
+
     @required
     eventStream: EventStream
 }
